LPD-15256 Move widgetPage and widgetPagesTest up, so they can be used by all test projects
